Refactor test/helper.sh and test/utils.go
* Move test/utils.go to test/utils/utils.go, and update corresponding
  references.
* Break test/helper.sh into separate functions in test/helpers, and
  source them from test/helper.sh.
* Files containing helper functions are explicitly sourced in
  test/helper.sh.
